requirements.txt
langchain
torch
sentence_transformers
faiss-cpu
huggingface-hub
pypdf
ctransformers
tiktoken
streamlit
streamlit-chat
psycopg2
flask